Make native _ckernels mandatory; consolidate solver backend
Reorganizes core solver components: moves KleinSolution/PerturbationSolution and klein_solve() into new solver_backend.py, removes optional numba fallbacks from CompiledModel (objective_vector_func, measurement_vector_func), adds native residual_eval() kernel for single-point equation evaluation, and creates tests/_oracles/core.py to house reference implementations for parity testing. The native C extension is now required rather than optional.
