+++ SymbolicDSGE/_ckernels/kalman/__init__.py
@@ -5,5 +5,7 @@ this module raises ``ImportError`` and the consumer (``SymbolicDSGE.kalman``)
 falls back to its numba kernels.
 """
 
-from ._kalman import kalman_hot_loop as kalman_hot_loop
-from ._kalman import ukf_hot_loop as ukf_hot_loop
+from ._kalman import kalman_hot_loop
+from ._kalman import ukf_hot_loop
+
+__all__ = ["kalman_hot_loop", "ukf_hot_loop"]
