+++ pyproject.toml
@@ -95,6 +95,7 @@ include = ["SymbolicDSGE*"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+pythonpath = ["tests"]
 python_files = ["test_*.py"]
 addopts = "--import-mode=importlib"
 markers = [
