+++ tests/core/test_solver_compiled.py
@@ -97,25 +97,6 @@ def test_compiled_equations_reject_bad_parameter_vector_length(compiled_test):
         c.equations(np.zeros(n), np.zeros(n), np.zeros(len(c.calib_params) - 1))
 
 
-def test_construct_measurement_vector_func_returns_expected_length(compiled_test):
-    c = compiled_test
-    f = c.construct_measurement_vector_func()
-    n = len(c.cur_syms)
-    p = len(c.calib_params)
-    args = [float64(0.0)] * (n + p)
-
-    out = f(*args)
-    assert out.shape == (len(c.observable_names),)
-
-
-def test_construct_measurement_vector_func_is_cached(compiled_test):
-    c = compiled_test
-
-    assert (
-        c.construct_measurement_vector_func() is c.construct_measurement_vector_func()
-    )
-
-
 def test_construct_measurement_array_dispatchers_are_cached(compiled_test):
     c = compiled_test
     obs = list(c.observable_names)
@@ -128,45 +109,6 @@ def test_construct_measurement_array_dispatchers_are_cached(compiled_test):
     ) is c.construct_observable_jacobian_array_func(obs)
 
 
-def test_construct_objective_vector_func_is_cached(compiled_test):
-    c = compiled_test
-
-    assert c.construct_objective_vector_func() is c.construct_objective_vector_func()
-
-
-def test_objective_vector_func_matches_compiled_equations(compiled_test):
-    c = compiled_test
-    objective = c.construct_objective_vector_func()
-    n = len(c.var_names)
-    fwd = np.linspace(0.1, 0.1 * n, n, dtype=np.complex128)
-    cur = np.linspace(-0.05, 0.05, n, dtype=np.complex128)
-    params = np.array(
-        [float64(c.config.calibration.parameters[p]) for p in c.calib_params],
-        dtype=np.complex128,
-    )
-
-    expected = c.equations(fwd, cur, params)
-    actual = objective(
-        np.ascontiguousarray(fwd),
-        np.ascontiguousarray(cur),
-        np.ascontiguousarray(params),
-    )
-    assert np.allclose(actual, expected)
-
-    fwd_step = fwd.copy()
-    cur_step = cur.copy()
-    fwd_step[0] += 1e-30j
-    cur_step[-1] -= 2e-30j
-
-    expected_step = c.equations(fwd_step, cur_step, params)
-    actual_step = objective(
-        np.ascontiguousarray(fwd_step),
-        np.ascontiguousarray(cur_step),
-        np.ascontiguousarray(params),
-    )
-    assert np.allclose(actual_step, expected_step)
-
-
 def test_measurement_array_dispatchers_match_scalar_dispatchers(compiled_test):
     c = compiled_test
     state = np.linspace(0.05, 0.05 * len(c.cur_syms), len(c.cur_syms), dtype=float64)
@@ -176,7 +118,7 @@ def test_measurement_array_dispatchers_match_scalar_dispatchers(compiled_test):
     )
 
     scalar_measure = np.asarray(
-        c.construct_measurement_vector_func()(*state, *params), dtype=float64
+        [fn(*state, *params) for fn in c.observable_funcs], dtype=float64
     )
     array_measure_func = c.construct_measurement_array_func(c.observable_names)
     array_measure = array_measure_func(state, params)
